Fix flatpak builds
...after
<https://src.fedoraproject.org/rpms/java-11-openjdk/c/6eee73b250484d6c740ce27eb03f36afe0526775>
"Update to jdk-11.0.16.1+1" added the TestTranslations.java "test to ensure
timezones can be translated":  Similar to the previous
<https://src.fedoraproject.org/rpms/java-11-openjdk/c/1ac4052b443719e5c582eff9ce30691e043be01f>
"Fix flatpak builds", during a flatpak build of java-11-openjdk its
.../images/jdk/lib/tzdb.dat is a dangling symlink to
/app/share/javazi-1.8/tzdb.dat (but which will be a working symlink in at least
the assembled LibreOffice flatpak).  That causes execution of
TestTranslations.java during the build to fail due to a
java.io.FileNotFoundException when trying to access that tzdb.dat.  The easiest
fix appears to be to just not run that specific test for a flatpak build.